reorganizando o codigo das peças
